\name{q4rcc-package}
\alias{q4rcc-package}
\alias{q4rcc}
\docType{package}
\title{
\packageTitle{q4rcc}
}
\description{
\packageDescription{q4rcc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{q4rcc}
\packageIndices{q4rcc}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{q4rcc}

Maintainer: \packageMaintainer{q4rcc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
